test: implement device manager mock

+++ src/TwinCAT.Mdp/DataTypes/ModuleInfo.cs
@@ -6,7 +6,7 @@ namespace TwinCAT.Mdp.DataTypes
 {
 	public class ModuleInfo
 	{
-		public uint ID;
+		public ushort ID;
 		public ModuleType Type;
 	}
 }
